LIBWEB-6487: Fix thumbnail and title rendering for search results
Drupal seems to do some preprocessing on fields that can be multi-valued
like the object__title__display and iiif_thumbnail_sequence__uris

If the field is a single value, Drupal makes it a string, so we just
need to check first before we render.

Also removing the language tag from the titles and then joining
them with a pipe, as that was the convention in Archelon.

https://umd-dit.atlassian.net/browse/LIBWEB-6487