LIBWEB-6487: Update configuration for the custom search result element
Instead of adding the fields within the search result block, have the
configuration for the fields in the settings for the custom element in
the Search API endpoint, under Result Mappings.

Added two additional fields to generate the anchor tag to link to the
results page: id and base_path.

https://umd-dit.atlassian.net/browse/LIBWEB-6487