LIBWEB-6498: Updates to match LIT conventions
Rather than construct a string with html attributes and use unsafeHTML,
we can just create a list of html templates, and when we provide that
to the render, LIT will handle extracting the elements from the list.

Ultimately we'll end up with the same result!
Lit docs: https://lit.dev/docs/templates/lists/

Also to make what types we're working with from the settings more
clear, I created an interface for the fields. This helps with the
syntax highlighting.

https://umd-dit.atlassian.net/browse/LIBWEB-6498